refactored post request
